+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   in Mock build environments (#163).
 - Add reference to build task ID and instance name in RPM/deb package changelog
   entries (#15).
+- Add result exportable field to all runnable tasks.
 - cli:
   - Possibility to list artifacts in registries in remote instance with the REST
     API (#141)